Modified the active admin pieces just a bit

+++ app/admin/school_admins.rb
@@ -11,7 +11,7 @@ ActiveAdmin.register SchoolAdmin do
     end
   end
   action_item do
-    if current_page?(:action => 'show')
+    if current_page?(:action => 'show') && !school_admin.district_guid.present?
       link_to "Delete School Admin", resource_path(resource), :confirm => 'Are you sure?', :method => :delete
     end
   end
