+++ app/admin/students.rb
@@ -11,7 +11,7 @@ ActiveAdmin.register Student do
     end
   end
   action_item do
-    if current_page?(:action => 'show')
+    if current_page?(:action => 'show') && !student.district_guid.present?
       link_to "Delete Student", resource_path(resource), :confirm => 'Are you sure?', :method => :delete
     end
   end
